Purging unused component files to cleanup charts. Refactored MaltreatmentDetails out of plot. Renaming componnts to better align with UI and terminology of the project.
