--- client/src/components/ProTx/components/charts/MaltreatmentChart.css
@@ -1,1 +1,0 @@
-/* .maltreatment-chart{} */
