--- client/src/components/ProTx/components/charts/ObservedFeaturesChart.css
@@ -1,1 +1,0 @@
-/* .observed-features-chart{} */
