--- client/src/components/ProTx/components/charts/ObservedFeaturesPlot.css
@@ -1,85 +1,0 @@
-.data-error-message {
-  margin: 1rem;
-  text-align: center;
-  font-weight: 700;
-  color: var(--color-blood-orange);
-}
-.loading-spinner {
-  margin-top: 40%;
-}
-.observed-features-plot {
-  font-family: var(--font-type-roboto-condensed);
-  width: 100%;
-}
-.observed-features-plot-layout {
-  margin: 0 0.5vw;
-}
-.observed-features-plot-info {
-  border-bottom: 1px solid var(--color-gray-very-dark);
-}
-.observed-features-plot-info-region {
-  min-height: 4.5rem;
-  display: flex;
-}
-.observed-features-plot-selected-region-label,
-.observed-features-plot-selected-feature-label {
-  font-weight: var(--font-weight-normal-plus);
-  color: var(--color-gray);
-}
-.observed-features-plot-selected-region {
-  display: inline;
-  margin: 1vh;
-}
-.observed-features-plot-selected-region {
-  width: 100%;
-  text-align: center;
-}
-.observed-features-plot-selected-region-label {
-  font-size: var(--font-size-sm);
-  font-weight: var(--font-weight-normal-plus);
-}
-.observed-features-plot-selected {
-  text-align: center;
-  border-bottom: 1px solid var(--color-gray-very-dark);
-}
-.observed-features-plot-selected-feature {
-  margin: 0.5vh;
-}
-.observed-features-plot-selected-feature-value {
-  margin-left: 1rem;
-  font-size: calc(var(--font-size-normal) * 1.2);
-  font-weight: var(--font-weight-x-heavy);
-}
-.observed-features-plot-selected-region-value {
-  margin: 0 1vw;
-  font-size: calc(var(--font-size-normal) * 2.2);
-  font-weight: var(--font-weight-x-heavy);
-}
-.observed-features-plot-selected-region-code {
-  font-size: var(--font-size-sm);
-  font-style: italic;
-}
-.observed-feature-selection-label {
-  font-weight: var(--font-weight-normal-plus);
-}
-.observed-features-plot-info-summary {
-  margin: 0.8vh 1vw 0.5vh;
-  text-align: center;
-  font-size: var(--font-size-xs);
-  line-height: 2vh;
-  font-style: italic;
-}
-.observed-features-plot-selected-type-value,
-.observed-features-plot-selected-type {
-  font-weight: var(--font-weight-heavy);
-}
-.observed-features-plot-selected-type-summary {
-  font-weight: var(--font-weight-x-heavy);
-}
-.observed-features-plot-selected-type-summary::after {
-  font-weight: var(--font-weight-normal);
-  content: '.';
-}
-.observed-features-plot-chart {
-  margin: 0 0.5vw;
-}
